fix: add missing return type to stopApp for linter compliance
